packages/debrief-backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==1.10.8

# Environment and utilities
python-dotenv==1.0.1

# Optional: for production deployment
gunicorn==21.2.0

# AI
openai==2.6.1